refactor(transaction-pay-controller): cleanup Polymarket relay integration
- Move polymarketRelayerUrl feature flag under payStrategies.relay so it
  sits with the other relay-strategy settings instead of at the
  feature-flag root.
- Collapse the verbose JSDoc on isPolymarketDepositWallet to the
  single-line shape matching the existing two instances.
- Remove the redundant if/else split in executeSingleQuote: the
  nonce-removal step now runs unconditionally before any source-leg
  branch, then a single if-else-if-else dispatches to hyperliquid,
  polymarket, or the default submitTransactions path.
- Fold wallet-batch-typed-data into relayer-api.ts and drop the
  PolymarketRelayerApi class. Replaced with standalone utils getNonce
  and submitDepositWalletBatch that take the messenger directly,
  resolve the URL via getPolymarketRelayerUrl, build the typed data,
  call KeyringController:signTypedMessage, and poll the relayer to a
  terminal state in one call.
- Inline the four-line applyPolymarketDepositWalletOverrides into
  withdraw.ts so the polymarket subdirectory drops the quotes.ts file
  entirely; the override applier sits next to the withdraw helper that
  uses the resulting quote.
- One-line CHANGELOG entry now that the per-step detail belongs to the
  PR description not the changelog.
